better UBO invalidation
Only invalidate UBOs when the data actually changes. It turns out it's
a fairly common case for application to update MaterialInstance
parameters with the same value; we now check for this before
marking the UBO "dirty".

The overhead of this check should be very small but it could save
many very expensive buffer updates.